backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.0.2
certifi==2022.9.24
click==8.1.3
click-didyoumean==0.3.0
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.3
django-appconf==1.0.5
django-cacheops==5.1
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-fsm==2.7.1
django-health-check==3.16.2
django-imagekit==4.1.0
django-jazzmin==2.6.0
django-json-widget==1.0.1
django-redis==5.2.0
djangorestframework==3.14.0
factory-boy==3.1.0
Faker==15.3.4
funcy==1.17
future==0.18.2
ipython==7.9.0
jedi==0.18.2
kombu==5.2.4
mysqlclient==2.1.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.3.0
prompt-toolkit==2.0.10
psutil==5.8.0
psycopg2==2.9.5
ptyprocess==0.7.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.4
sentry-sdk==0.19.2
six==1.16.0
sqlparse==0.4.3
traitlets==5.6.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
